pyproject.toml
[project]
name = "ziddfit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.6",
    "django==6.0.1",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-auth>=2.49.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt[crypto]>=2.12.1",
    "razorpay>=2.0.1",
    "requests>=2.33.0",
]
